requirements.txt
ultralytics>=8.0.0
mediapipe>=0.10.0
opencv-contrib-python>=4.8.0
torch>=2.0.0
boxmot
supervision>=0.16.0
numpy>=1.24.0
pyyaml>=6.0
torchvision
